chore: refactor email code
For now, the code to send emails is commented out and the dependency on sendgrid is removed.
